refactor: extract regenerate-plugin-tocode
