+++ scripts/character-class-escape-sets.js
@@ -3,6 +3,7 @@
 const fs = require('fs');
 const jsesc = require('jsesc');
 const regenerate = require('regenerate');
+require('./utils/regenerate-plugin-to-code.js');
 
 const Zs = require('@unicode/unicode-16.0.0/General_Category/Space_Separator/code-points.js');
 
@@ -82,38 +83,6 @@ addCharacterClassEscape(
 );
 
 /*----------------------------------------------------------------------------*/
-
-const codePointToString = (codePoint) => {
-	return '0x' + codePoint.toString(16).toUpperCase();
-};
-
-// Regenerate plugin that turns a set into some JavaScript source code that
-// generates that set.
-regenerate.prototype.toCode = function() {
-	const data = this.data;
-	// Iterate over the data per `(start, end)` pair.
-	let index = 0;
-	let start;
-	let end;
-	const length = data.length;
-	const loneCodePoints = [];
-	const ranges = [];
-	while (index < length) {
-		start = data[index];
-		end = data[index + 1] - 1; // Note: the `- 1` makes `end` inclusive.
-		if (start == end) {
-			loneCodePoints.push(codePointToString(start));
-		} else {
-			ranges.push(
-				'addRange(' + codePointToString(start) +
-				', ' + codePointToString(end) + ')'
-			);
-		}
-		index += 2;
-	}
-	return 'regenerate(' + loneCodePoints.join(', ') + ')' +
-		(ranges.length ? '\n\t\t.' + ranges.join('\n\t\t.') : '');
-};
 
 const stringify = (name, object) => {
 	const source = 'exports.' + name + ' = new Map([\n\t' + Object.keys(object).map((character) => {
